localrequirements.txt
absl-py==2.3.1
addict==2.4.0
anyio==4.9.0
attrs==25.3.0
autolab-core==1.1.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dill==0.4.0
exceptiongroup==1.3.0
filelock==3.18.0
fonttools==4.59.0
freetype-py==2.5.1
fsspec==2025.7.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
lxml==6.0.0
manifold3d==3.2.0
mapbox_earcut==1.0.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.4.2
numpy==1.26.4
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
protobuf==6.31.1
pycollada==0.9.2
pyglet==2.1.6
PyOpenGL==3.1.0
pyparsing==3.2.3
pyrender==0.1.45
python-dateutil==2.9.0.post0
pyzed @ file:///usr/local/zed/pyzed-5.0-cp310-cp310-linux_aarch64.whl
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rtree==1.4.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
setproctitle==1.3.6
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
svg.path==7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
torch @ file:///home/mulengangulale/Downloads/torch-2.8.0-cp310-cp310-linux_aarch64.whl
torchaudio @ file:///home/mulengangulale/Downloads/torchaudio-2.8.0-cp310-cp310-linux_aarch64.whl
torchvision @ file:///home/mulengangulale/Downloads/torchvision-0.23.0-cp310-cp310-linux_aarch64.whl
tqdm==4.67.1
trimesh==4.7.1
typing_extensions==4.14.1
urllib3==2.5.0
vhacdx==0.0.8.post2
visualization==1.0.0
Werkzeug==3.1.3
xxhash==3.5.0
